+++ src/idc/writer/imgseg/_indexedpng.py
@@ -12,7 +12,7 @@ from idc.api import ImageSegmentationData, SplittableStreamWriter, make_list, de
 class IndexedPngImageSegmentationWriter(SplittableStreamWriter, AnnotationsOnlyWriter):
 
     def __init__(self, output_dir: str = None,
-                 image_path_rel: str = None, palette: str = None, annotations_only: bool = None,
+                 image_path_rel: str = None, palette: str = None, annotations_only: bool = None, background: int = None,
                  split_names: List[str] = None, split_ratios: List[int] = None,
                  logger_name: str = None, logging_level: str = LOGGING_WARNING):
         """
@@ -26,6 +26,8 @@ class IndexedPngImageSegmentationWriter(SplittableStreamWriter, AnnotationsOnlyW
         :type palette: str
         :param annotations_only: whether to output only the annotations and not the images
         :type annotations_only: bool
+        :param background: the index (0-255) to use as background
+        :type background: int
         :param split_names: the names of the splits, no splitting if None
         :type split_names: list
         :param split_ratios: the integer ratios of the splits (must sum up to 100)
@@ -40,6 +42,7 @@ class IndexedPngImageSegmentationWriter(SplittableStreamWriter, AnnotationsOnlyW
         self.image_path_rel = image_path_rel
         self.palette = palette
         self.annotations_only = annotations_only
+        self.background = background
         self._palette_list = None
 
     def name(self) -> str:
@@ -71,6 +74,7 @@ class IndexedPngImageSegmentationWriter(SplittableStreamWriter, AnnotationsOnlyW
         parser.add_argument("-o", "--output", type=str, help="The directory to store the images files in. Any defined splits get added beneath there.", required=True)
         parser.add_argument("--image_path_rel", metavar="PATH", type=str, default=None, help="The relative path from the annotations to the images directory", required=False)
         parser.add_argument("-p", "--palette", metavar="PALETTE", type=str, default=PALETTE_AUTO, help="The palette to use; either palette name (%s) or comma-separated list of R,G,B values." % "|".join(PALETTES), required=False)
+        parser.add_argument("--background", type=int, help="The index (0-255) to use for the background", required=False, default=0)
         add_annotations_only_param(parser)
         return parser
 
@@ -86,6 +90,7 @@ class IndexedPngImageSegmentationWriter(SplittableStreamWriter, AnnotationsOnlyW
         self.image_path_rel = ns.image_path_rel
         self.palette = ns.palette
         self.annotations_only = ns.annotations_only
+        self.background = ns.background
 
     def accepts(self) -> List:
         """
@@ -153,6 +158,8 @@ class IndexedPngImageSegmentationWriter(SplittableStreamWriter, AnnotationsOnlyW
                         sub_arr = item.annotation.layers[label]
                         sub_arr = np.where(sub_arr == 255, index, 0).astype(np.uint8)
                         arr += sub_arr
+                if self.background > 0:
+                    arr = np.where(arr == 0, self.background, arr)
                 ann = Image.fromarray(arr, "P")
                 ann.putpalette(self._palette_list)
                 path = sub_dir